Fix compilation under 64-bit linux

+++ dsl_details.cc
@@ -1052,8 +1052,8 @@ void expandTildes( wstring & str, wstring const & tildeReplacement )
       if( x > 0 && str[ x - 1 ] == '^' && ( x < 2 || str[ x - 2 ] != '\\' ) )
       {
         str.replace( x - 1, 2, tildeReplacement );
-        str[ x - 1 ] = QChar( str[ x - 1 ] ).isUpper() ? QChar::toLower( str[ x - 1 ] )
-                                                       : QChar::toUpper( str[ x - 1 ] );
+        str[ x - 1 ] = QChar( str[ x - 1 ] ).isUpper() ? QChar::toLower( (uint)str[ x - 1 ] )
+                                                       : QChar::toUpper( (uint)str[ x - 1 ] );
         x = x - 1 + tildeReplacement.size();
       }
       else
